havok_serde

Module ser

Serialization

Re-exports

pub use std::error::Error as StdError;

Enums

TypeSize

Used for writing binary data in Array. (To write after the pointer type data)

Traits

Error

Trait used by Serialize implementations to generically construct errors belonging to the Serializer against which they are currently running.

Serialize

A data structure that can be serialized into any data format supported by Serde.

SerializeFlags

Returned from Serializer::serialize_enum_flags.

SerializeSeq

Returned from Serializer::serialize_array.

SerializeStruct

Returned from Serializer::serialize_struct.

Serializer

A data format that can serialize any data structure supported by Serde.