havok_serde::ser
Trait Error

pub trait Error: Sized + StdError {
    // Required method
    fn custom<T>(msg: T) -> Self
       where T: Display;
}

Trait used by Serialize implementations to generically construct errors belonging to the Serializer against which they are currently running.

Required Methods

fn custom<T>(msg: T) -> Self
where T: Display,

Used when a Serialize implementation encounters any error while serializing a type.

The message should not be capitalized and should not end with a period.

For example, a filesystem Path may refuse to serialize itself if it contains invalid UTF-8 data.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.