havok_serde::ser

Trait Serialize

pub trait Serialize {
    // Required method
    fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error>;
}

A data structure that can be serialized into any data format supported by Serde.

Serde provides Serialize implementations for all C++ Havok types.

Additionally, Serde provides a procedural macro called [havok_serde_derive] to automatically generate Serialize implementations for structs and enums in your program. See the [derive section of the manual] for how to use this.

In rare cases it may be necessary to implement Serialize manually for some type in your program. See the [Implementing Serialize] section of the manual for more about this.

Support types

Unsupported types

There’re never used in the Havok classes.(ver. hk2010)

Zero, FunctionPointer, InplaceArray, HomogeneousArray, RelArray, Max

Required Methods

fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error>

Serialize this value into the given Serde serializer.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.