havok_serde

Module de

Deserialization

Re-exports

pub use std::error::Error as StdError;

Enums

ReadEnumSize

Size to read to deserialize enum and flags. This is used to deserialize binary data

Unexpected

Unexpected represents an unexpected invocation of any one of the Visitor trait methods.

Traits

ClassIndexAccess

Class Indexed Deserializer trait

Deserialize

A data structure that can be deserialized from any data format supported by Serde.

DeserializeOwned

A data structure that can be deserialized without borrowing any data from the deserializer.

DeserializeSeed

DeserializeSeed is the stateful form of the Deserialize trait. If you ever find yourself looking for a way to pass data into a Deserialize impl, this trait is the way to do it.

Deserializer

Deserializer

EnumAccess

Provides a Visitor access to the data of an enum in the input.

Error

The Error trait allows Deserialize implementations to create descriptive error messages belonging to the Deserializer against which they are currently running.

Expected

Expected represents an explanation of what data a Visitor was expecting to receive.

MapAccess

Provides a Visitor access to each entry of a map in the input.

SeqAccess

Provides a Visitor access to each element of a sequence in the input.

VariantAccess

VariantAccess is a visitor that is created by the Deserializer and passed to the Deserialize to deserialize the content of a particular enum variant.

Visitor

This trait represents a visitor that walks through a deserializer.